CRM:行程记录打卡成功返回新插入数据的rptVstAuto

+++ provider/provider-order-api/src/main/java/com/funtl/myshop/plus/provider/api/RptVstService.java
@@ -68,5 +68,5 @@ public interface RptVstService{
      * @param rptVst
      * @return
      */
-    Integer insert(RptVst rptVst);
+    Long insert(RptVst rptVst);
 }
